Moved pagination concerns out of dashboards namespace
